transaction_verification/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.6
watchdog==6.0.0
geopy==2.4.1
dnspython==2.7.0